Respond to review feedback
