+++ components/provider_ppucd/Cargo.toml
@@ -25,7 +25,6 @@ include = [
 icu_locid = { version = "0.1", path = "../locid" }
 icu_provider = { version = "0.1", path = "../provider" }
 serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
 icu_locid_macros = { version = "0.1", path = "../locid/macros" }
 tinystr = "0.4"
 icu_uniset = { version = "0.1", path = "../uniset" }
